requirements.txt
bentoml==1.4.23
joblib==1.5.2
torch==2.8.0
torchvision==0.23.0
torchsummary==1.5.1
tqdm==4.67.1
numpy==2.2.6
matplotlib==3.10.6
opencv-python==4.12.0.88
pillow==11.3.0
pandas==2.3.2
seaborn==0.13.2
streamlit==1.49.1